Kernel/IntelGraphics: Prepare to support future Display plane formats
The Intel Skylake register set for display planes is a bit different, so
let's ensure we can properly support it in the near future.
