+++ Kernel/Graphics/Intel/DisplayConnectorGroup.h
@@ -10,9 +10,9 @@
 #include <AK/Try.h>
 #include <Kernel/Graphics/Console/GenericFramebufferConsole.h>
 #include <Kernel/Graphics/Intel/Definitions.h>
-#include <Kernel/Graphics/Intel/DisplayPlane.h>
 #include <Kernel/Graphics/Intel/GMBusConnector.h>
 #include <Kernel/Graphics/Intel/NativeDisplayConnector.h>
+#include <Kernel/Graphics/Intel/Plane/DisplayPlane.h>
 #include <Kernel/Graphics/Intel/Transcoder/DisplayTranscoder.h>
 #include <Kernel/Library/LockRefPtr.h>
 #include <Kernel/Memory/TypedMapping.h>
