+++ Kernel/Graphics/Intel/Plane/DisplayPlane.h
@@ -38,14 +38,16 @@ public:
 public:
     static ErrorOr<NonnullOwnPtr<IntelDisplayPlane>> create_with_physical_address(PhysicalAddress plane_registers_start_address);
 
-    ErrorOr<void> set_plane_settings(Badge<IntelDisplayConnectorGroup>, PhysicalAddress aperture_start, PipeSelect, size_t horizontal_active_pixels_count);
+    virtual ErrorOr<void> set_plane_settings(Badge<IntelDisplayConnectorGroup>, PhysicalAddress aperture_start, PipeSelect, size_t horizontal_active_pixels_count) = 0;
     ErrorOr<void> enable(Badge<IntelDisplayConnectorGroup>);
     bool is_enabled(Badge<IntelDisplayConnectorGroup>);
     ErrorOr<void> disable(Badge<IntelDisplayConnectorGroup>);
 
     ShadowRegisters shadow_registers() const;
 
-private:
+    virtual ~IntelDisplayPlane() = default;
+
+protected:
     struct [[gnu::packed]] PlaneRegisters {
         u32 control;
         u32 linear_offset;
